Add outgroup option
Add an option to include an outgroup taxon inthe beast configuration.
This is sometimes useful for including the original founder sequence in
a phylogenetic tree.  Now that we will be allowing indels in the
generated sequences (so simulated sequences are not all the same
length), we will have to include outgroups prior to doing a multiple
alignment, so this option is probably not useful any more.